final changes made for the day
Need to work on input. You can hit notes just fine, but if there are two notes that are hittable at the same time, the logic that I have written only acknowledges one. I should store notes that are hittable within an array and pass them down the index in order. Hopefully that works!